requirements.txt
streamlit>=1.30
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
plotly>=5.18
sqlalchemy>=2.0
scipy>=1.11
statsmodels>=0.14
xgboost>=2.0
openpyxl>=3.1
PyPDF2>=3.0
python-docx>=1.1
